Change view functions to non-view as required to keep the rates cache up to date

+++ src/products/interfaces/IRateOracle.sol
@@ -46,6 +46,6 @@ interface IRateOracle {
         uint256 beforeTimestamp,
         UD60x18 atOrAfterIndex,
         uint256 atOrAfterTimestamp,
-        uint40 queryTimestamp
+        uint256 queryTimestamp
     ) external pure returns (UD60x18 interpolatedIndex);
 }
